FillMetadata: Stay away from Category tags
..because these are already filled in during Import

In Import we know from linguisitic info which pages are categories, in
this stage we don't have this info anymore. The previous solution was to
derive the universe of categories from all pages in this CBOR. This is
brittle and would not allow partial conversion of one archive at a time.